Fix incorrect quota usage calculation

+++ features/merchant/seller_tokopedia_flash_sale/src/main/java/com/tokopedia/tkpd/flashsale/presentation/list/child/FlashSaleListViewModel.kt
@@ -101,7 +101,7 @@ class FlashSaleListViewModel @Inject constructor(
 
     private fun FlashSale.findQuotaUsagePercentage() : Int {
         val usedQuota = maxProductSubmission - remainingQuota
-        return (usedQuota / maxProductSubmission) * PERCENT
+        return ((usedQuota.toFloat() / maxProductSubmission.toFloat()) * PERCENT).toInt()
     }
 
     private fun FlashSale.toOngoingItem() : DelegateAdapterItem {
